backend/requirements.txt
Django 

djangorestframework 

django-cors-headers 

djangorestframework-simplejwt 

daphne 

Pillow 

python-dotenv 

django-filter 

django-environ 

forex-python 

environ 

Twisted[tls,http2] 